Separate out auth controller.
Also fix mistaken deletion of show exceptions after handler line.
